feat(ds): OtpField gains invalid, describedBy and testId; split the sanitiser (#474)
Three props the migration proved were missing. A control used outside a
Field has no way to say it is in error or to point at an external error
message, and both MFA dialogs are exactly that: they label the code
themselves and carry their own error line. testId keeps the existing
data-testid hooks on the two auth fields.

sanitiseCode moves to its own module for the reason fieldContext.ts
exists — a module exporting both a component and a plain function cannot
be hot-replaced, and react-refresh/only-export-components is at error
here. It needs a second consumer anyway: the MFA LOGIN field cannot be an
OtpField (it accepts a 12-character recovery code) but needs exactly this
cleaning.
