+++ frontend/src/shared/ds/index.ts
@@ -48,7 +48,8 @@ export { Tooltip, type TooltipProps, type TooltipPlacement } from './Tooltip';
 export { Popover, type PopoverProps, type PopoverPlacement } from './Popover';
 export { Menu, type MenuProps, type MenuItem } from './Menu';
 export { Command, type CommandProps, type CommandItem } from './Command';
-export { OtpField, type OtpFieldProps, type OtpAlphabet } from './OtpField';
+export { OtpField, type OtpFieldProps } from './OtpField';
+export { sanitiseCode, type OtpAlphabet } from './otpCode';
 export {
   ErrorState,
   PermissionDenied,
